+++ Pixen Application/Animation/PXAnimationDocument.m
@@ -107,7 +107,7 @@
 		OSProgressPopup *popup = [OSProgressPopup sharedProgressPopup];
 		PXAnimatedGifExporter *exporter = [[[PXAnimatedGifExporter alloc] initWithSize:[animation size] iterations:1] autorelease];
 		int i;
-		int numberOfCels = [animation countOfCels];
+		NSUInteger numberOfCels = [animation countOfCels];
 		[popup setMaxProgress:numberOfCels];
 		[popup beginOperationWithStatusText:[NSString stringWithFormat:@"Exporting GIF... (1 of %d)", numberOfCels] parentWindow:[windowController window]];
 		[popup setProgress:0];
